requirements.txt
Pillow==10.3.0
Requests==2.32.3
streamlit==1.36.0
streamlit_lottie==0.0.5
